Add new directories and files for project
I added new files and directories to this new
project. I added an "index.html" file to the root
directory. I added two new directories within the
root directory; an "images" directory and a
"recipes" directory respectively. Within the
"recipes" directory I am going to use the
"taco.html" file as a template of this project
since it outlines all requirements by The Odin
Project. This will make the other two recipes
much easier to complete in a shorter amount of
time. In the "images" directory I saved images
to use in their respective recipes pages.